Use lint-staged for formatting
