+++ .lintstagedrc.yml
@@ -1,0 +1,1 @@
+'!(js,ts,jsx,tsx)': prettier --write --ignore-unknown
